add make.ps1

+++ make.sh
@@ -9,44 +9,50 @@ Options:
 EOF
 )
 
-function priv_use
+function priv_lazbuild
 (
-    [[ -d 'use' ]] || mkdir 'use'
-    for item in "${@}"; do
-        declare -A VAR=(
-            [url]="https://packages.lazarus-ide.org/${item}.zip"
-            [out]=$(mktemp)
-            [dir]=${item##/*}
-        )
-        wget --output-document "${VAR[out]}" "${VAR[url]}"
-        unzip -o "${VAR[out]}" -d "use/${VAR[dir]}"
-        rm --verbose "${VAR[out]}"
-    done
-)
-
-function pub_build
-(
-    git submodule update --init --recursive
-    priv_use 'PlaysoundPackage' 'Rx' 'FPSpreadsheet'
-    find 'use' -type 'f' -name '*.lpk' -exec lazbuild --add-package-link {} \;
-    find 'src' -type 'f' -name '*.lpi' -exec lazbuild --recursive --build-mode=release {} \;
-)
-
-function priv_main
-(
-    set -euo pipefail
     if ! (which lazbuild); then
         source '/etc/os-release'
         case ${ID:?} in
             debian | ubuntu)
                 sudo apt-get update
                 sudo apt-get install -y lazarus
-            ;;
+                ;;
         esac
     fi
+    declare -r COMPONENTS='use/components.txt'
+    if [[ -d "${COMPONENTS%%/*}" ]]; then
+        git submodule update --init --recursive --force --remote
+        if [[ -f "${COMPONENTS}" ]]; then
+            while read -r; do
+                if [[ -n "${REPLY}" ]] &&
+                    ! (lazbuild --verbose-pkgsearch "${REPLY}") &&
+                    ! (lazbuild --add-package "${REPLY}") &&
+                    ! [[ -d "${COMPONENTS%%/*}/${REPLY}" ]]; then
+                        declare -A VAR=(
+                            [url]="https://packages.lazarus-ide.org/${REPLY}.zip"
+                            [out]=$(mktemp)
+                        )
+                        wget --output-document "${VAR[out]}" "${VAR[url]}" >/dev/null
+                        unzip -o "${VAR[out]}" -d "${COMPONENTS%%/*}/${REPLY}"
+                        rm --verbose "${VAR[out]}"
+                    fi
+            done < "${COMPONENTS}"
+        fi
+        find "${COMPONENTS%%/*}" -type 'f' -name '*.lpk' -exec \
+            lazbuild --add-package-link {} +
+    fi
+    find 'src' -type 'f' -name '*.lpi' -exec \
+        lazbuild --no-write-project --recursive --no-write-project --build-mode=release {} + 1>&2
+)
+
+function priv_main
+(
+    set -euo pipefail
     if ((${#})); then
         case ${1} in
-            build) pub_build 1>&2 ;;
+            build) priv_lazbuild ;;
+            *) priv_clippit ;;
         esac
     else
         priv_clippit
